touranement page completed

+++ src/app/components/tournaments/tournaments.component.ts
@@ -3,11 +3,14 @@ import { TournamentService } from '../../services/tournament.service';
 import { Observable } from 'rxjs/Observable';
 import { Router } from '@angular/router';
 import { AuthService } from '../../services/auth.service';
+import { NgbDropdownConfig } from '@ng-bootstrap/ng-bootstrap';
+import swal from 'sweetalert2';
 
 @Component({
   selector: 'app-tournaments',
   templateUrl: './tournaments.component.html',
-  styleUrls: ['./tournaments.component.css']
+  styleUrls: ['./tournaments.component.css'],
+  providers: [NgbDropdownConfig]
 })
 export class TournamentsComponent implements OnInit {
   listAllTournaments: Observable<object[]>;
@@ -16,9 +19,11 @@ export class TournamentsComponent implements OnInit {
 
   constructor(private tournamentService: TournamentService,
               private router: Router,
-              private authService: AuthService) 
+              private authService: AuthService,
+              private config: NgbDropdownConfig) 
               {
                 this.userStatus = this.authService.status;
+                config.placement = 'right-top';
               }
               
   ngOnInit() { 
@@ -26,6 +31,7 @@ export class TournamentsComponent implements OnInit {
     if (this.userStatus.status === "manager") {
       this.tournamentService.listAllTournamentsForManager(this.userStatus.team_id)
       this.listAllTournamentsForManager = this.tournamentService.tournamentForManager;
+      this.listAllTournamentsForManager.subscribe(res => console.log(res))
     } 
     // if user is organizer / player
     else {
@@ -34,8 +40,34 @@ export class TournamentsComponent implements OnInit {
   }
 
   // delete tournament
-  onDelete(id: number) {
-    this.tournamentService.delete(id);
+  async onDelete(id: number) {
+    
+    const result = await this.swalSetUp();
+
+    if (result.value) {
+      this.tournamentService.delete(id);
+      swal({
+        type: 'success',
+        confirmButtonText: 'DELETED!',
+        width: 300,
+      })
+    } else if (result.dismiss === swal.DismissReason.cancel) {
+      swal({
+        type: 'error',
+        confirmButtonText: 'Canceled',
+        width: 300,
+      })
+    }    
+  }
+
+  swalSetUp() {
+    return swal({
+      title: 'Are you sure?',
+      type: 'info',
+      showCancelButton: true,
+      confirmButtonText: 'Yes',
+      cancelButtonText: 'Cancel'
+    })
   }
 
   joinTournament(tournamentID) {
